UI update, can select from a UI list of guides.
With the new UI update:

An additional guides window appears with a list of guides that we can multi-select.

We now have a self.bipedal_guides list of strings for the guide types that updates both the guide UI and then passes this data to the functions to generate.

Tested that we can now add in non-limb guide types to JSON.